% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{area_data}
\alias{area_data}
\title{tissue core area data.}
\format{
A data.frame with 200 rows (core_ids in row.names) and 3 columns:
\describe{
  \item{S}{stromal area in mm2, positive numeric values}
  \item{T}{tumor intraepitheliao area in mm2, positive numeric values}
  \item{overall}{overall area in mm2, positive numeric values}
  }
}
\usage{
area_data
}
\description{
A data.frame containing containing 200 rows for 100 cores comprising 100 cases/tumors
}
\keyword{datasets}
