% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pheno_density.r
\name{calculate_pheno_density}
\alias{calculate_pheno_density}
\title{Compute cell phenotype density}
\usage{
calculate_pheno_density(
  area_data = NULL,
  input_dir = NULL,
  output_dir = NULL,
  pheno_fnm = "cell_phenotype.RData"
)
}
\arguments{
\item{area_data}{A data.frame containing the area data (e.g. mm2) of each core (in rows, core_ids are
in row.names); the tissue categories (i.e. intraepithelial, stroma, overall) are represented in columns.
Note that matching annotation of tissue categories should be used in \code{area_data} and \code{phenotype_data}
identified in Step 3 previously. Note that "overall" naming is hard coded.}

\item{input_dir}{A file.path string of input directory storing the PIPA phenotype assignment of all cells of
interest}

\item{output_dir}{A file.path string of output directory}

\item{pheno_fnm}{A string indicating .RData object filename of the PIPA phenotype assignment}
}
\description{
Compute cell phenotype density
}
