% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{CCC}
\alias{CCC}
\title{Cell consensus clustering function}
\usage{
CCC(
  data = NULL,
  batch_size = 3000,
  consensus_rep = 5,
  max_k = 20,
  output_dir = getwd(),
  phenograph_k = 100,
  max_batch = NULL
)
}
\arguments{
\item{data}{Input data: samples in rows, features in columns; cell_ids in row.names}

\item{batch_size}{An integer indicating the cell window size i.e. number of cells in each batch run}

\item{consensus_rep}{An integer indicating the number of repeats for consensus clustering (per window/batch run)}

\item{max_k}{An integer indicating the maximum number of clusters used in hierachical clustering (i.e. k in cutree)
of the consensus matrix, at each bath run}

\item{output_dir}{A file.path string of output directory}

\item{phenograph_k}{An integer indicating the number of neighbors in PhenoGraph}

\item{max_batch}{An integer indicating the maximum number of windows/batchs}
}
\description{
Cell consensus clustering function using PhenoGraph to group cells
}
