% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{cell_data}
\alias{cell_data}
\title{Cell feature data.}
\format{
A data frame with 10000 rows and 3 identity columns + 19 cell feature columns: \describe{
  \item{feature1}{cell feature 1, numeric values} \item{feature2}{cell feature 2, numeric values}
  \item{feature3}{cell feature 3, numeric values} \item{feature4}{cell feature 4, numeric values}
  \item{feature5}{cell feature 5, numeric values} \item{feature6}{cell feature 6, numeric values}
  \item{feature7}{cell feature 7, numeric values} \item{feature8}{cell feature 8, numeric values}
  \item{feature9}{cell feature 9, numeric values} \item{feature10}{cell feature 10, numeric values}
  \item{feature11}{cell feature 11, numeric values} \item{feature12}{cell feature 12, numeric values}
  \item{feature13}{cell feature 13, numeric values} \item{feature14}{cell feature 14, numeric values}
  \item{feature15}{cell feature 15, numeric values} \item{feature16}{cell feature 16, numeric values}
  \item{feature17}{cell feature 17, numeric values} \item{feature18}{cell feature 18, numeric values}
  \item{feature19}{cell feature 19, numeric values}
  \item{cell_ids}{cell identifiers with XY locations, character strings}
  \item{tumor_ids}{tumor or patient identifiers, character strings}
  \item{core_ids}{core identifiers, character strings} }
}
\usage{
cell_data
}
\description{
A dataset containing 20000 cell of interest,
15793 cells in stromal region and 4207 cells in tumor region
}
\keyword{datasets}
