% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_phenotype_assignment.r
\name{cell_phenotype_assignment}
\alias{cell_phenotype_assignment}
\title{Cell phenotype assignment}
\usage{
cell_phenotype_assignment(
  cell_data = NULL,
  log10_featureNms = NULL,
  selected_feature_dir = NULL,
  knn_neighbors = 100,
  seed = 999,
  model_fnm = "final_meta_clusters_by_prog.RData",
  cell_model_input_dir = NULL,
  output_dir = NULL,
  save_intermediate = FALSE,
  min_fraction = 0.001
)
}
\arguments{
\item{cell_data}{A data.frame contains cells in rows and features in columns; cell_ids in row.names
the column 'tumor_ids' are the tumor identifiers; the column 'core_ids' are the core identifiers;
the column 'cell_ids' are the cell identifiers containing core_ids+coordinates}

\item{log10_featureNms}{A vector of characters of feature names whose values to be log10-transformed prior to scaling (z-score)}

\item{selected_feature_dir}{A file.path string of input directory for selected features
(output from {meta_cluster})}

\item{knn_neighbors}{An integers of the number of neighbors to be used \code{kNN}}

\item{seed}{An integer of seed for reproducibility}

\item{model_fnm}{A character string indicating the file name for the R object results output from
\code{meta_cluster}}

\item{cell_model_input_dir}{A file.path string of output directory that the cell model (i.e. meta cluster assignment)
to be loaded from}

\item{output_dir}{A file.path string of output directory that the resultant all cell phenotype assignment is saved}

\item{save_intermediate}{A boolean indicates the saving option for cell assignment of individual cell batches}

\item{min_fraction}{A fraction (0 to 1) indicating the minimum abundance of phenotype (among all cells of
interest) to be included in the differential analysis (default: 0.01 i.e. 1 percent)}
}
\description{
Cell phenotype assignment for all cell of interest, using kNN
}
