% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_surv_analysis.r
\name{composition_surv_analysis}
\alias{composition_surv_analysis}
\title{Post-PIPA analysis - prognostic association analysis based on cell phenotype composition of tumors}
\usage{
composition_surv_analysis(
  surv_data = NULL,
  phenotype_fnm = NULL,
  density_data = NULL,
  phenotype_dir = NULL,
  output_dir = NULL,
  seed = 999,
  min_cluster_size = 20,
  neighbor_size_byFrac = c(0.05, 0.1),
  KM_xlab = "Survivals",
  area_split = TRUE
)
}
\arguments{
\item{surv_data}{A data.frame: tumors in rows (tumor_ids in row.names), columns must include: "time", "cens";
if there are additional columns, they are treated as confounders}

\item{phenotype_fnm}{A character string of input file name for cell phenotype data}

\item{density_data}{A (long format) data.frame containing cell density data (for heatmap viz):
samples/tumors in rows with 4 columns named as: "phenotype", "area", "tumor_ids", "density"}

\item{phenotype_dir}{A file.path string of input directory storing the PIPA phenotype assignment of all cells of
interest}

\item{output_dir}{A file.path string of output directory}

\item{seed}{An integer of seed for reproducibility}

\item{min_cluster_size}{An integer indicating the minimum cluster size i.e. cluster/tumor subtypes containing
fewer tumors than this number will be excluded in downstream survival analysis
i.e. input argument for \code{surv_analysis_fn} \code{Viz_heatmap_bySubtype}}

\item{neighbor_size_byFrac}{A numerical vector, in fraction, indicating the fraction of total samples/tumors to be
used in PhenoGraph}

\item{KM_xlab}{A character string for labelling the x-axis in Kaplan-Meier curves}

\item{area_split}{A boolean indicating if the analysis (fraction calculations) will be split by area}
}
\description{
Post-PIPA analysis - prognostic association analysis based on cell phenotype composition of tumors
Tumor subtyping is performed using \link[Rphenograph]{Rphenograph}, cell phenotype composition of PP,IP&GP as input
}
