% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aggregated_feature.r
\name{compute_aggregated_feature}
\alias{compute_aggregated_feature}
\title{Compute patient/tumor-level aggregated cell features}
\usage{
compute_aggregated_feature(data = NULL, output_dir = getwd())
}
\arguments{
\item{data}{A data.frame contains cells in rows and features in columns;
the column 'tumor_ids' represents the tumor identifiers}

\item{output_dir}{A file.path string of output directory}
}
\value{
A data.frame contains tumor-level feature summaries saved in
  \code{output_dir}
}
\description{
Compute the min, max, 25th, 50th, 75th, and stdev of each cellular features across all tumors/cases
}
