% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{connectivityMatrix}
\alias{connectivityMatrix}
\title{Consensus matrix updating function}
\usage{
connectivityMatrix(clusterAssignments, m, sampleKey)
}
\arguments{
\item{clusterAssignments}{A named vector of length X-samples containing the input cluster assignment of each sample}

\item{m}{A data.frame of size X-by-X containing the cumulated consensus cluster assignment}

\item{sampleKey}{A vector of length X containing the sample identifiers}
}
\description{
Consensus matrix updating function
}
