% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection_extremeProgGps.r
\name{featureSelection_extremeProgGps}
\alias{featureSelection_extremeProgGps}
\title{Feature selection - identify and evaluate patient groups with extreme prognoses}
\usage{
featureSelection_extremeProgGps(
  surv_data = NULL,
  subtype = NULL,
  save_res = TRUE,
  output_dir = NULL,
  out_bnm = "FS_extremeProgGps",
  min_cluster_size = 20
)
}
\arguments{
\item{surv_data}{A data.frame: tumors in rows (tumor_ids in row.names), columns must include: "time", "cens";
if there are additional columns, they are treated as confounders}

\item{subtype}{A data.frame representing the tumor subtype data; it contains 2 columns named as tumor_ids
and cluster_no}

\item{save_res}{A boolean indicating if the results to be saved}

\item{output_dir}{A file.path string of output directory}

\item{out_bnm}{A character string for naming the output result in .RData}

\item{min_cluster_size}{An integer indicating the minimum cluster size i.e. cluster/tumor subtypes containing
fewer tumors than this number will be excluded in the survival analysis}
}
\description{
Feature selection - identify and evaluate patient groups with extreme prognoses.
(1) identifying patient groups (among all output from \code{featureSelection_patientGrouping}) with extreme prognoses
i.e., most unfavorable vs. favorable
(2) computing concordance of Kaplan-Meier estimates as a measure of noise & separation
}
\details{
## step (2) at each unique subset of features tested in step(1),
}
