% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection_ProgSigFeature.r
\name{featureSelection_ProgSigFeature}
\alias{featureSelection_ProgSigFeature}
\title{Prognostic feature selection using Cox PH model}
\usage{
featureSelection_ProgSigFeature(
  feature_data = NULL,
  surv_data = NULL,
  log10_featureNms = NULL,
  regularization_alpha = 1,
  output_dir = NULL,
  no_of_runs = 10
)
}
\arguments{
\item{feature_data}{A data.frame contains tumors in rows and aggregated cellular features in columns;
row.names represent the tumor identifiers}

\item{surv_data}{A data.frame contains tumors in rows and survival data in columns:
"cens" for censoring status; "time" for surviving time; row.names represent the tumor identifiers}

\item{log10_featureNms}{A vector of characters of feature names for log10 tranformation prior to z-score
transformation; all other features are transformed to z-scores}

\item{regularization_alpha}{A fraction (0 to 1) representing the alpha value
for \link[glmnet]{cv.glmnet}}

\item{output_dir}{A file.path indicating the output directory}

\item{no_of_runs}{An integer of the number of repeats of surviving test}
}
\value{
A data.frame contains tumor-level feature summaries saved in
  \code{output_dir}
}
\description{
Prognostic feature selection using Cox PH model with regularization methods
(Ridge with alpha=0; Lasso with alpha=1;ElasticNet with alpha in the range 0,1)
}
