% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection_wrapper.r
\name{featureSelection_wrapper}
\alias{featureSelection_wrapper}
\title{Feature selection wrapper function}
\usage{
featureSelection_wrapper(
  data = NULL,
  surv_data = NULL,
  log10_featureNms = NULL,
  no_of_runs = 10,
  output_dir = NULL,
  min_cluster_size = 20
)
}
\arguments{
\item{data}{A data.frame contains cells in rows and features in columns;
the column 'tumor_ids' represents the tumor identifiers
input argument for \code{compute_aggregated_feature}}

\item{surv_data}{A data.frame: tumors in rows (tumor_ids as row.names), columns must include: "time", "cens";
"cens" for censoring status; "time" for surviving time; row.names represent the tumor identifiers;
if there are additional columns, they are treated as confounders and will not be used for feature selection
input argument for \code{featureSelection_ProgSigFeature} & \code{featureSelection_extremeProgGps}}

\item{log10_featureNms}{A vector of characters of feature names for log10 tranformation prior to z-score
transformation; all other features are transformed to z-scores;
input argument for \code{featureSelection_ProgSigFeature}}

\item{no_of_runs}{An integer of the number of repeats of surviving test;
input argument for \code{featureSelection_ProgSigFeature}}

\item{output_dir}{A file.path string of output directory}

\item{min_cluster_size}{An integer indicating the minimum cluster size i.e. cluster/tumor subtypes containing
fewer tumors than this number will be excluded in the survival analysis
input argument for \code{featureSelection_extremeProgGps}}
}
\description{
Feature selection wrapper function: step (1) compute_aggregated_feature,
step (2) featureSelection_ProgSigFeature, step (3) featureSelection_patientGrouping
step (4) featureSelection_extremeProgGps,
step (5) make visualizatin plots for selecting the optimal subset of cell features
}
