% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{forest_plot}
\alias{forest_plot}
\title{Forest plot summarizing Cox PH regression analysis results}
\usage{
forest_plot(
  HR_CI = NULL,
  output_dir = NULL,
  param_bnm = "test",
  tissue_areas = c("S", "T"),
  plot_width = 8,
  plot_height = 8
)
}
\arguments{
\item{HR_CI}{A data.frame containing the HR and CI data}

\item{output_dir}{A file.path string of output directory}

\item{param_bnm}{A character string of the identifier of the experimental setting of interest
to be used in \link[ggplot2]{facet_wrap}; and it will appended to the PDF output filename}

\item{tissue_areas}{A vector of characters indicating the tissue regions in which density quartiles
are computed are to be used for making forest plots; it has to be consistent with the tissue region
naming found under the 'area' column in input data HR_CI}

\item{plot_width}{An integer to be passed to the 'width' argument in
\link[ggplot2]{ggsave}}

\item{plot_height}{An integer to be passed to the 'height' argument in
\link[ggplot2]{ggsave}}
}
\value{
Number PDF files equal to the length of tissue_areas
}
\description{
Forest plot summarizing Cox PH regression analysis results - supporting function for \code{ForesPlot_denQ_CoxPH}
}
