% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{log10_transform}
\alias{log10_transform}
\title{Data transformation function}
\usage{
log10_transform(data = NULL, method = "log10")
}
\arguments{
\item{data}{Input data: samples (cells or tumors) in rows, features in columns}

\item{method}{A character string indicate the type of data transformation
(current version does only 1og10 transformation)}
}
\description{
Data transformation function: log10; log(value <= 0) assigned to log(min. value >0)
}
