% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{merge_CCC}
\alias{merge_CCC}
\title{Meta-clustering of the cell-consensus-clusters}
\usage{
merge_CCC(data = NULL)
}
\arguments{
\item{data}{A list whose elments are clustering assignment data.frame output from \code{CCC}}
}
\description{
Meta-clustering of the cell-consensus-clusters obtained from overlapping data windows
}
