% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cluster.r
\name{meta_cluster}
\alias{meta_cluster}
\title{Construct meta_cluster}
\usage{
meta_cluster(
  cell_data = NULL,
  log10_featureNms = NULL,
  batch_size = 3000,
  input_dir = NULL,
  output_dir = NULL,
  prognosis_delineation = TRUE,
  cell_batch_clustering = TRUE,
  cell_batch_merging = TRUE,
  User_ClusterNo = NULL,
  max_batch = NULL,
  phenograph_k = 100,
  consensus_rep = 5,
  remove_cor_features = FALSE,
  no_of_random_merging = 100,
  seed = 999
)
}
\arguments{
\item{cell_data}{A data.frame contains cells in rows and features in columns; cell_ids in row.names
the column 'tumor_ids' are the tumor identifiers; the column 'core_ids' are the core identifiers;
the column 'cell_ids' are the cell identifiers containing core_ids+coordinates}

\item{log10_featureNms}{A vector of characters of feature names whose values to be log10-transformed prior to scaling (z-score)}

\item{batch_size}{An integer indicating the cell window size i.e. number of cells in each batch run;
to be passed to \code{CCC}}

\item{input_dir}{A file.path string of input directory}

\item{output_dir}{A file.path string of output directory that the resultant phenotyping model is saved}

\item{prognosis_delineation}{A boolean indicating if the resultant phenotypes to be delineated by tumor prognosis
subtypes (as identified in Step 1)}

\item{cell_batch_clustering}{A boolean: TRUE (Step 2) cell clustering in batches will be performed; FALSE
previous results will be loaded for merging}

\item{cell_batch_merging}{A boolean: TRUE (Step 2) the merging of cell clusters from individual batches
will be performed;}

\item{User_ClusterNo}{An integer: User-specified number of meta clusters}

\item{max_batch}{An integer indicating the maximum number of windows/batchs used in \code{CCC}}

\item{phenograph_k}{An integer indicating the number of neighbors in PhenoGraph used in \code{CCC}}

\item{consensus_rep}{An integer indicating the number of repeats for consensus clustering (per window/batch run);
to be passed to \code{CCC}}

\item{remove_cor_features}{A boolean indicating if highly correlated features should be removed}

\item{no_of_random_merging}{An integer indicating the number merging with different (randomized) ordering of
cell windows; for identifying the inherent no. of phenotypes}

\item{seed}{An integer for reproducibility}
}
\description{
Construct meta_cluster using training cells
}
