% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize_Params.r
\name{Optimize_Params}
\alias{Optimize_Params}
\title{Optimize parameters}
\usage{
Optimize_Params(
  root_dir = NULL,
  pheno_dir_path = NULL,
  output_dir = NULL,
  param_bnm = NULL,
  pheno_dens_fnm = "cell_phenotype_density.RData",
  pheno_count_fnm = "cell_phenotype.RData",
  tissue_area_list = c("T", "S"),
  plot_bnm = "",
  plot_width = 8,
  plot_height = 8
)
}
\arguments{
\item{root_dir}{A file.path pointing to the parent directory of the folders containing
results using different param settings}

\item{pheno_dir_path}{A character string of the directory path to the (final) phenotype & density data}

\item{output_dir}{A file.path string of output directory}

\item{param_bnm}{A character string of the key name for identifying the folder names of
the experimental setting of interest}

\item{pheno_dens_fnm}{A character string of input file name for cell phenotype density data
(after merging i.e. PP, IP, GP)}

\item{pheno_count_fnm}{A character string of input file name for cell phenotype count data
(after merging i.e. PP, IP, GP)}

\item{tissue_area_list}{A character vectors indicating the tissue areas of interest for plotting}

\item{plot_bnm}{A character string identifier to be appended to the output PDF}

\item{plot_width}{An integer to be passed to the 'width' argument in
\link[ggplot2]{facet_wrap}}

\item{plot_height}{An integer to be passed to the 'height' argument in
\link[ggplot2]{ggsave}}
}
\value{
Two PDF files
}
\description{
Compare and optimize parameters, based on the distribution of (a) %cell, and
(b) prevalence of patient/tumor for  PP, IP, and GP phenotypes
}
