% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{patient_data}
\alias{patient_data}
\title{Clinical outcome and clinicopathological data.}
\format{
A data frame containing 100 rows (i.e. cases/tumors) and 6 columns:
  \describe{ \item{time}{survival time, numeric} \item{cens}{censoring data,
  binary} \item{cov1}{covariate var1, factor}
  \item{cov2}{covariate var, factor} \item{cov3}{covariate var, factor}
  \item{cov4}{covariate var, factor}}
}
\usage{
patient_data
}
\description{
A data frame containing the survival and clinicopathological data for 100 tumors.
}
\keyword{datasets}
