% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPA-package.R
\docType{package}
\name{PIPA-package}
\alias{PIPA}
\alias{PIPA-package}
\title{PIPA: Prognosis-informed Phenotype Assignment (PIPA)}
\description{
Prognosis-informed Phenotype Assignment (PIPA): A Novel Method for Unsupervised Discovery of Cell Phenotypes with Prognostic Significance in the Tumor Microenvironment.
}
\author{
\strong{Maintainer}: MaiChan Lau \email{lau_mai_chan@bii.a-star.edu.sg}

Authors:
\itemize{
  \item Jonathan Nowak \email{janowak@bwh.harvard.edu}
  \item Shuji Ogino \email{sogino@bwh.harvard.edu}
}

}
\keyword{internal}
