% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{raw_dens_data}
\alias{raw_dens_data}
\title{Tumor denisty data.}
\format{
A data.frame with 200 rows and 4 columns:
\describe{
  \item{tumor_ids}{tumor or patient identifiers, character strings}
  \item{area}{S for stromal and T for tumor intraepithelial region, character strings}
  \item{phenotype}{cell phenotype name, character strings}
  \item{density}{density values, positive numeric values}
  }
}
\usage{
raw_dens_data
}
\description{
A data.frame containing containing 200 rows, 100 each in stromal and
tumor area (i.e. 100 tumors)
}
\keyword{datasets}
