% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_pheno_byProg.r
\name{refine_pheno_byProg}
\alias{refine_pheno_byProg}
\title{Refine phenotypes through merging PP or GP phenotypes based on prognostic significance}
\usage{
refine_pheno_byProg(
  surv_data = NULL,
  output_dir = NULL,
  KM_xlab = "Survivals",
  preMerge_density_dir = NULL,
  preMerge_density_fnm = "cell_phenotype_density.RData",
  preMerge_pheno_dir = NULL,
  preMerge_pheno_fnm = "cell_phenotype.RData",
  uni_Cox_dir = NULL,
  uni_Cox_fnm = "uniCoxPH_summary.txt",
  uni_P_cutoff = 0.005
)
}
\arguments{
\item{surv_data}{A data.frame: tumors in rows (tumor_ids in row.names), columns must include: "time", "cens";
if there are additional columns, they are treated as confounders}

\item{output_dir}{A file.path string of output directory}

\item{KM_xlab}{A character string for labelling the x-axis in Kaplan-Meier curves}

\item{preMerge_density_dir}{A file.path string of input directory storing the phenotype cell density data
before merging}

\item{preMerge_density_fnm}{A character string of input file name for cell density data before merging}

\item{preMerge_pheno_dir}{A file.path string of input directory storing the cell phenotype data
before merging}

\item{preMerge_pheno_fnm}{A character string of input file name for cell phenotype data before merging}

\item{uni_Cox_dir}{A file.path string of input directory storing the univariable Cox PH analysis data}

\item{uni_Cox_fnm}{A character string of input file name for univariable Cox PH analysis data}

\item{uni_P_cutoff}{A numeric value indicating the p-value cutoff for selecting phenotypes}
}
\description{
Refine phenotypes through merging PP or GP phenotypes based on prognostic significance
(Step 1) Keep phenotypes show Ptrend < uni_P_cutoff in univariable Cox PH analysis
(Step 2) Keep phenotypes show association (i.e., non-zero coefficient) in a multivariable Cox
PH model which includes phenotypes remains from Step 1; separate models for PP and GP phenotypes.
}
