% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{surv_analysis_fn}
\alias{surv_analysis_fn}
\title{Survival analysis wrapper function}
\usage{
surv_analysis_fn(
  surv_data = NULL,
  subtype = NULL,
  user_ref = NULL,
  output_dir = NULL,
  plot_width = 8,
  out_bnm = "CoxPH",
  min_cluster_size = 20,
  KM_xlab = "Survivals"
)
}
\arguments{
\item{surv_data}{A data.frame: tumors in rows (tumor_ids in row.names), columns must include: "time", "cens";
if there are additional columns, they are treated as confounders}

\item{subtype}{A data.frame representing the tumor subtype data; it contains 2 columns named as tumor_ids
and cluster_no}

\item{user_ref}{A character indicating the reference group (ie. tumor subtype)}

\item{output_dir}{A file.path string of output directory}

\item{plot_width}{An integer of the PDF file width}

\item{out_bnm}{A character string for naming the output PDF file}

\item{min_cluster_size}{An integer indicating the minimum cluster size i.e. cluster/tumor subtypes containing
fewer tumors than this number will be excluded in the survival analysis}

\item{KM_xlab}{A character string for labelling the x-axis in Kaplan-Meier curves}
}
\description{
Survival analysis wrapper function
}
