% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_feature_boxplot_byPheno.r
\name{Viz_feature_boxplot_byPheno}
\alias{Viz_feature_boxplot_byPheno}
\title{Box-plot visualization: distribution of cell features grouped by PIPA-derived phenotypes}
\usage{
Viz_feature_boxplot_byPheno(
  cell_data = NULL,
  phenotype_fnm = NULL,
  pheno_dir = NULL,
  output_dir = NULL,
  plot_bnm = "PIPA_features",
  pheno_levels = c("PP", "IP", "GP"),
  outlier_removal = TRUE,
  ncol = 1,
  plot_width = 8,
  plot_height = 8
)
}
\arguments{
\item{cell_data}{A data.frame contains cells in rows and features in columns; MUST contain a
column named 'cell_ids' as cell identifiers (core_ids & XY-coordinates) and other columns
are treated as features for plotting}

\item{phenotype_fnm}{A character string of input file name for cell phenotype data}

\item{pheno_dir}{A file.path string of input directory storing the PIPA phenotype assignment
of all cells of interest}

\item{output_dir}{A file.path string of output directory}

\item{plot_bnm}{A character string identifier appended to the output folder name}

\item{pheno_levels}{A vector of characters represent the phenotype levels (i.e., ordering of plotting)}

\item{outlier_removal}{A boolean indicating if outliers (outside IQR) are included}

\item{ncol}{An integer indicating the number of columns (ncol argument) in
\link[ggplot2]{facet_wrap}}

\item{plot_width}{An integer to be passed to the 'width' argument in
\link[ggplot2]{facet_wrap}}

\item{plot_height}{An integer to be passed to the 'height' argument in
\link[ggplot2]{ggsave}}
}
\value{
A PDF file containing the box-plot
}
\description{
Box-plot visualization: distribution of cell features grouped by PIPA-derived phenotypes
}
