% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_feature_heatmap_byPheno.r
\name{Viz_feature_heatmap_byPheno}
\alias{Viz_feature_heatmap_byPheno}
\title{Heat-map visualization: distribution of cell features grouped by PIPA-derived phenotypes}
\usage{
Viz_feature_heatmap_byPheno(
  cell_data = NULL,
  phenotype_fnm = "cell_phenotype.RData",
  pheno_dir = NULL,
  output_dir = NULL,
  pheno_levels = c("PP", "IP", "GP"),
  column_scale = TRUE,
  filetype = "pdf",
  plot_width = NULL,
  plot_height = 8
)
}
\arguments{
\item{cell_data}{A data.frame contains cells in rows and features in columns; MUST contain a column named 'cell_ids'
as cell identifiers (core_ids+coordinates) and a column named 'tumor_ids' for tumor/case identifiers}

\item{phenotype_fnm}{A character string of input file name for cell phenotype data}

\item{pheno_dir}{A file.path string of input directory storing the PIPA phenotype assignment of all cells of
interest}

\item{output_dir}{A file.path to the output directory}

\item{pheno_levels}{A vector of characters represent the phenotype levels (i.e., ordering of plotting)}

\item{column_scale}{A boolean indicating if heatmap values are column/marker-wise scaled}

\item{filetype}{A character string indicating the output plot file format, pdf or png}

\item{plot_width}{An integer to be passed to the 'width' argument in
\link[ggplot2]{ggsave}}

\item{plot_height}{An integer to be passed to the 'height' argument in
\link[ggplot2]{ggsave}}
}
\value{
PDF & PNG files containing the heat maps
}
\description{
Heat-map visualization: distribution of cell features grouped by PIPA-derived phenotypes
using (1) case/tumor-median feature values, and (2) cellular feature values of
PIPA-selected features
}
