% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_feature_UMAP.R
\name{Viz_feature_UMAP}
\alias{Viz_feature_UMAP}
\title{UMAP visualization: relatedness of cells in the PIPA-selected feature space}
\usage{
Viz_feature_UMAP(
  cell_data = NULL,
  downspl_size = 1000,
  phenotype_fnm = "cell_phenotype.RData",
  seed = 123,
  log10_featureNms = NULL,
  pheno_dir = NULL,
  output_dir = NULL,
  plot_width = 8,
  plot_height = 8
)
}
\arguments{
\item{cell_data}{A data.frame contains cells in rows and features in columns; MUST contain a column named 'cell_ids'
as cell identifiers (core_ids+coordinates) and all other columns will be used as cell features for UMAP dimensional
reduction}

\item{downspl_size}{An integer indicating the number of randomly selected cells
(downspl_size cells from PP, IP and GP individually) to be used for UMAP dimensional reduction}

\item{phenotype_fnm}{A character string of input file name for cell phenotype data}

\item{seed}{An integer used as the seed for generating the downsampled cells}

\item{log10_featureNms}{A vector of characters of feature names whose values to be log10-transformed
prior to scaling (z-score); same setting as \code{meta_cluster} should be used}

\item{pheno_dir}{A file.path string of input directory storing the PIPA phenotype assignment of all cells of
interest}

\item{output_dir}{A file.path to the output directory}

\item{plot_width}{An integer to be passed to the 'width' argument in
\link[ggplot2]{ggsave}}

\item{plot_height}{An integer to be passed to the 'height' argument in
\link[ggplot2]{ggsave}}
}
\value{
PDF files containing the UMAP plots
}
\description{
UMAP visualization: relatedness of cells in the PIPA-selected feature space - annotation by
(1) PIPA-derived phenotypes, and (2) cellular feature values of PIPA-selected features
}
