% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_foresPlot_denQ_CoxPH.r
\name{Viz_foresPlot_denQ_CoxPH}
\alias{Viz_foresPlot_denQ_CoxPH}
\title{Forest plot summarizing Cox PH regression analysis results}
\usage{
Viz_foresPlot_denQ_CoxPH(
  root_dir = NULL,
  output_dir = root_dir,
  pheno_dir_path = NULL,
  Cox_result_folderNm = "density_surv/cat",
  uni_fnm = "uniCoxPH_",
  multi_fnm = "multiCoxPH_",
  param_bnm = NULL
)
}
\arguments{
\item{root_dir}{A file.path pointing to the parent directory of the folders containing
results of multiple runs using different setting (to be examined)}

\item{output_dir}{A file.path string of output directory}

\item{pheno_dir_path}{A character string of the directory path to the (final) phenotype folder}

\item{Cox_result_folderNm}{A character string of the directory path to the folder containing Cox PH output
results generated by \code{density_surv_analysis}}

\item{uni_fnm}{A character string used to search for the univariable Cox PH results
(as filename prefix) output by \code{density_surv_analysis}}

\item{multi_fnm}{A character string used to search for the multivariable Cox PH results
(as filename prefix) output by \code{density_surv_analysis}}

\item{param_bnm}{A character string of the key name for identifying the folder names of
the experimental setting of interest}
}
\value{
PDF files containing forest plot
}
\description{
Forest plot summarizing Cox PH regression analysis results (univariable or multivariable models)
using tumor subtypes of density quartiles, across different param settings e.g. detection rates for feature selection
or batch_size/ max_batch/ no_of_random_merging in \code{meta_cluster}
}
