% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_heatmap_bySubtype.r
\name{Viz_heatmap_bySubtype}
\alias{Viz_heatmap_bySubtype}
\title{Heat-map visualization: distribution of features grouped by tumor subtypes}
\usage{
Viz_heatmap_bySubtype(
  data = NULL,
  density_data = NULL,
  subtype = NULL,
  output_dir = NULL,
  min_cluster_size = 20,
  log10_dens = TRUE,
  value_name = "Value",
  row_title = "Feature",
  feature_order = NULL,
  plot_bnm = NULL,
  column_scale = FALSE,
  plot_width = 12,
  plot_height = 10
)
}
\arguments{
\item{data}{A data.frame containing tumor/case-level values:
tumors/cases in rows with 1+N columns named as: "tumor_ids","value1", "value2" etc for N features}

\item{density_data}{A (long format) data.frame containing cell density data:
tumors/cases in rows with 4 columns named as: "phenotype", "area", "tumor_ids", "density"}

\item{subtype}{A data.frame containing tumor subtyping data:
tumors/cases in rows with 2 columns named as: "tumor_ids","cluster_no"}

\item{output_dir}{A file.path to the output directory}

\item{min_cluster_size}{An integer indicating the minimum cluster size (i.e. no. of cases/tumors) for which
a tumor subtype containing tumors less than this number will be excluded in the heatmap#'}

\item{log10_dens}{A boolean: 1 if density is log10 transformed, 0 if otherwise}

\item{value_name}{A character indicating the name argument passed to \link[ComplexHeatmap]{Heatmap}}

\item{row_title}{A character indicating the row_title argument passed to \link[ComplexHeatmap]{Heatmap}}

\item{feature_order}{A vector of characters represent the feature levels (i.e., ordering of plotting)}

\item{plot_bnm}{A character to be appended to the output plot filename}

\item{column_scale}{A boolean indicating if heatmap values are column/marker-wise scaled}

\item{plot_width}{An integer to be passed to the 'width' argument in
\link[ggplot2]{ggsave}}

\item{plot_height}{An integer to be passed to the 'height' argument in \link[ggplot2]{ggsave}}
}
\description{
Heat-map visualization: distribution of features (e.g. cellular features, PIPA-dervied phenotype density)
grouped by tumor subtypes with optional annotation of cell density data
}
